update functions, error in injecting data

+++ src/main/java/com/codecraft/agora_backend/repository/AdminEmailsRepository.java
@@ -3,5 +3,5 @@ package com.codecraft.agora_backend.repository;
 import com.codecraft.agora_backend.model.AdminEmails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminEmailsRepository extends JpaRepository<AdminEmails, Integer> {
+public interface AdminEmailsRepository extends JpaRepository<AdminEmails, Long> {
 }
